server/pyproject.toml
[tool.poetry]
name = "merch store"
version = "0.1.0"
description = ""
authors = ["Dino Perovic <dino.perovic@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "~4.1.7"
django-salesman = {extras = ["pygments"], version = "~1.1.6"}

[tool.poetry.group.dev.dependencies]
pyright = "^1.1.299"
black = "^23.1.0"
isort = "^5.12.0"
ruff = "^0.0.256"
django-stubs = "^1.16.0"

[tool.pyright]
typeCheckingMode = "strict"

[tool.isort]
profile = "black"

[tool.ruff]
select = [
  "F",
  "E",
  "W",
  "I",
  "C",
  "B",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
